Implements nuget/csproj parser
